Duplicated code fix
Home and project/audio_controller both have scroll logic that is the same. Creating 1 file for the scroll indicators and 1 file for the actual logic behind the scrolling
