+++ src/pages/home.jsx
@@ -1,51 +1,19 @@
-import React, { useRef, useState, useEffect } from "react";
+import React from "react";
 import ProgrammingIMG from '../img/programming.png';
 import ProjectCard from '../components/projectCards';
 import { programTitle, programText, loremIpsumTitle, loremIpsumText } from '../texts/texts';
+import ScrollIndicator from "../components/scrollIndicator";
+import UseScrollTracker from "../components/useScrollTracker";
 
 function Home() {
-  const containerRef = useRef(null);
-  const [scrollThumbTop, setScrollThumbTop] = useState(0);
-  const [scrollThumbHeight, setScrollThumbHeight] = useState(0);
-  const [currentSection, setCurrentSection] = useState(0);
-  const [sectionCount, setSectionCount] = useState(0);
-
-  useEffect(() => {
-    const handleResize = () => {
-      if (containerRef.current) {
-        const { clientHeight, scrollHeight } = containerRef.current;
-
-        // Dynamically count sections
-        const sections = containerRef.current.querySelectorAll(".section").length;
-        setSectionCount(sections);
-
-        // Calculate the thumb height
-        const thumbHeight = (clientHeight / scrollHeight) * 100;
-        setScrollThumbHeight(thumbHeight);
-      }
-    };
-
-    window.addEventListener("resize", handleResize);
-    handleResize();
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
-  const handleScroll = () => {
-    if (containerRef.current) {
-      const { scrollTop, clientHeight, scrollHeight } = containerRef.current;
-
-      // Update thumb position
-      const thumbTop = (scrollTop / (scrollHeight - clientHeight)) * (100 - scrollThumbHeight);
-      setScrollThumbTop(thumbTop);
-
-      // Determine the current section index
-      const current = Math.round(scrollTop / clientHeight);
-      setCurrentSection(current);
-    }
-  };
+  const {
+    containerRef,
+    scrollThumbTop,
+    scrollThumbHeight,
+    currentSection,
+    sectionCount,
+    handleScroll
+  } = UseScrollTracker();
 
   return (
     <>
@@ -74,23 +42,12 @@ function Home() {
             <h1>Section 3</h1>
           </section>
         </div>
-        <div className="custom-scrollbar">
-          <div
-            className="scroll-thumb"
-            style={{
-              height: `${scrollThumbHeight}%`,
-              top: `${scrollThumbTop}%`,
-            }}
-          ></div>
-        </div>
-        <div className="indicator">
-          {Array.from({ length: sectionCount }, (_, index) => (
-            <div
-              key={index}
-              className={`dot ${currentSection === index ? "active" : ""}`}
-            ></div>
-          ))}
-        </div>
+        <ScrollIndicator
+          scrollThumbHeight={scrollThumbHeight}
+          scrollThumbTop={scrollThumbTop}
+          sectionCount={sectionCount}
+          currentSection={currentSection}
+        />
       </div>
     </>
   )
